+++ component/src/main/java/arez/component/AbstractEntityReference.java
@@ -1,6 +1,5 @@
 package arez.component;
 
-import arez.Disposable;
 import arez.annotations.Observable;
 import arez.annotations.ObservableRef;
 import arez.annotations.PreDispose;
@@ -11,10 +10,9 @@ import javax.annotation.Nullable;
  * Abstract base class for reference to a component where reference should be cleared when the component is disposed.
  */
 public abstract class AbstractEntityReference<T>
-  extends AbstractEntryContainer<T>
 {
   @Nullable
-  private EntityEntry<T> _entry;
+  private T _entity;
 
   /**
    * Dispose or detach the associated entity if any.
@@ -22,7 +20,7 @@ public abstract class AbstractEntityReference<T>
   @PreDispose
   protected void preDispose()
   {
-    doSetEntity( null );
+    setEntity( null );
   }
 
   /**
@@ -33,7 +31,7 @@ public abstract class AbstractEntityReference<T>
   @Nullable
   protected T getEntity()
   {
-    return getEntityUnlessDisposed();
+    return _entity;
   }
 
   /**
@@ -44,20 +42,7 @@ public abstract class AbstractEntityReference<T>
   protected boolean hasEntity()
   {
     getEntityObservable().reportObserved();
-    return null != getEntityUnlessDisposed();
-  }
-
-  /**
-   * Return the entity if reference exists and entity is not disposed.
-   *
-   * @return the entity if it is not disposed.
-   */
-  @Nullable
-  private T getEntityUnlessDisposed()
-  {
-    return null == _entry ?
-           null :
-           Disposable.isNotDisposed( _entry.getEntity() ) ? _entry.getEntity() : null;
+    return null != _entity;
   }
 
   /**
@@ -77,31 +62,35 @@ public abstract class AbstractEntityReference<T>
   @Observable
   protected void setEntity( @Nullable final T entity )
   {
-    doSetEntity( entity );
+    if ( null != _entity )
+    {
+      detachEntity( _entity );
+    }
+    _entity = entity;
+    if ( null != _entity )
+    {
+      attachEntity( _entity );
+    }
   }
 
-  private void doSetEntity( @Nullable final T entity )
+  private void attachEntity( @Nonnull final T entity )
   {
-    if ( null != _entry )
-    {
-      detachEntry( _entry, false );
-    }
-    if ( null == entity )
-    {
-      _entry = null;
-    }
-    else
-    {
-      _entry = createEntityEntry( entity, reference -> {
-        /*
-         * Can not use setEntity( null ) here as the generated method will check equality
-         * prior to mutating. getEntity will return null as entity has been disposed so
-         * setter will not actually run.
-         */
+    DisposeTrackable
+      .asDisposeTrackable( entity )
+      .getNotifier()
+      .addOnDisposeListener( this, () -> {
         getEntityObservable().preReportChanged();
-        detachEntry( _entry, false );
+        detachEntity( entity );
+        _entity = null;
         getEntityObservable().reportChanged();
       } );
-    }
+  }
+
+  private void detachEntity( @Nonnull final T entity )
+  {
+    DisposeTrackable
+      .asDisposeTrackable( entity )
+      .getNotifier()
+      .removeOnDisposeListener( this );
   }
 }